convert some str to unicode + move get_matchdict functions
